refactor: extract serve core config and context modules
