+++ packages/cli/src/serve/core/config/aliases/sdcpp-video.ts
