+++ packages/cli/src/serve/core/fastify.d.ts
@@ -1,0 +1,17 @@
+import type { QvacContext, QvacRequestModel } from '@/serve/core/context'
+import type { ParsedFile } from '@/serve/lib/multipart'
+
+declare module 'fastify' {
+  interface FastifyInstance {
+    qvac: QvacContext
+  }
+  interface FastifyRequest {
+    qvacModel?: QvacRequestModel
+    bindCancel: (requestId: string) => void
+    multipartFiles?: ParsedFile[]
+  }
+  interface FastifyContextConfig {
+    unsupportedParams?: string[]
+    sseSentinel?: boolean
+  }
+}
