+++ packages/cli/src/serve/core/lifecycle.ts
@@ -1,5 +1,6 @@
 import { unloadModel as sdkUnloadModel, close as sdkClose } from '@qvac/sdk'
-import type { LoadConfig, ModelRegistry, ServeConfig } from '@/serve/core/model-registry'
+import type { ModelRegistry } from '@/serve/core/model-registry'
+import type { LoadConfig, ServeConfig } from '@/serve/core/config/types'
 import type { LoadManager } from '@/serve/core/load-manager'
 import type { Logger } from '@/logger'
