+++ packages/cli/src/serve/routes/models-catalog.ts
@@ -1,7 +1,7 @@
 import type { FastifyPluginAsyncZod } from 'fastify-type-provider-zod'
 import { HttpError } from '@/serve/lib/http-error'
 import { buildCatalog, filterCatalog, paginate } from '@/serve/core/model-catalog'
-import { loadModelConstants } from '@/serve/sdk-constants'
+import { loadModelConstants } from '@/serve/core/config/constants'
 import { modelCatalogQuery, modelCatalogIdParams } from '@/serve/schemas/models-catalog'
 
 const descriptions = {
