+++ packages/cli/test/unit/core/model-catalog.test.ts
@@ -2,7 +2,7 @@ import { describe, it } from 'node:test'
 import assert from 'node:assert/strict'
 import type { ModelConstant } from '@qvac/sdk'
 import { createModelRegistry } from '@/serve/core/model-registry'
-import type { ResolvedModelEntry, ServeConfig } from '@/serve/core/model-registry'
+import type { ResolvedModelEntry, ServeConfig } from '@/serve/core/config/types'
 import { buildCatalog, filterCatalog, paginate } from '@/serve/core/model-catalog'
 
 function constant(name: string, over: Partial<ModelConstant> = {}): ModelConstant {
